// Unimeet-Web-Client/src/app/typescript-angular-client/model/uniViewDto.ts
/**
 * Unimeet Service
 * User Api Documentation
 *
 * OpenAPI spec version: 1.0.0
 * Contact: emreecann28@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface UniViewDto { 
    uniID?: number;
    uniName?: string;
}